+++ Main/Main/Main.java
@@ -8,6 +8,12 @@ import javax.swing.SwingUtilities;
  */
 public class Main {
     public static void main(String[] args) {
+        try {
+            com.formdev.flatlaf.themes.FlatMacDarkLaf.setup();
+        } catch (Exception ex) {
+            System.err.println("Failed to initialize LaF");
+        }
+        
         SwingUtilities.invokeLater(() -> {
             new RoleSelectionFrame().setVisible(true);
         });